Assessment Error Handling / Track totals
Added functionality to track number of courses per term and number of assessments per course. Also fixed a bug that resulted in a course having an assessment count of less than 0. The fixes I made should eliminate that possibility, but I've added a couple of conditionals to default the total to 0 if it drops to a negative number.